bots/bq/requirements.txt
google-cloud-bigquery==3.22.0
re-ircbot==2.0.14-dev
python-dotenv==1.0.0
pandas==2.2.2
tabulate==0.9.0
